TEST_PROJECT/requirements.txt
django>=4.1.0,<5.0.0
djangorestframework
pyyaml 
requests
django-cors-headers
psycopg2-binary